# gin_contrib/example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "paths": {
        "/testapi/pet/get-pet-by-id/{pet_id}/": {
            "get": {
                "description": "get pet by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get a  pet",
                "operationId": "get-pet-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pet id",
                        "name": "pet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/testapi/update-product/{product_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Update product attributes",
                "operationId": "update-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "productInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProductUpdates"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "api.ProductUpdates": {
            "type": "object",
            "properties": {
                "description": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "stock": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "type": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}